+++ src/main/java/org/team4/shcParameters/ShcParameterController.java
@@ -31,7 +31,7 @@ public class ShcParameterController {
         windowChoiceBox.getItems().clear();
         doorsChoiceBox.getItems().clear();
 
-        if(!Settings.simulationStarted) {
+        if (!Settings.simulationStarted) {
             windowSetButton.setDisable(true);
             doorSetButton.setDisable(true);
             return;
@@ -41,16 +41,16 @@ public class ShcParameterController {
         doorSetButton.setDisable(false);
 
         String[] windowList = House.getAllWindowsOption();
-        for(int i = 0; i < windowList.length; i++) {
+        for (int i = 0; i < windowList.length; i++) {
             windowChoiceBox.getItems().add(windowList[i]);
         }
 
         String[] doorList = House.getAllDoorsOption();
-        for(int i = 0; i < doorList.length; i++) {
+        for (int i = 0; i < doorList.length; i++) {
             doorsChoiceBox.getItems().add(doorList[i]);
         }
-        if(windowList.length > 0) windowChoiceBox.setValue(windowList[0]);
-        if(doorList.length > 0) doorsChoiceBox.setValue(doorList[0]);
+        if (windowList.length > 0) windowChoiceBox.setValue(windowList[0]);
+        if (doorList.length > 0) doorsChoiceBox.setValue(doorList[0]);
     }
 
 
@@ -59,7 +59,7 @@ public class ShcParameterController {
      */
     public void toggleWindowAction() {
 
-        if(windowChoiceBox.getValue() != null) {
+        if (windowChoiceBox.getValue() != null) {
             House.toggleWindowOpen(windowChoiceBox.getValue());
             windowAndDoorChoiceBoxInit();
             DashboardController dashboardController = App.fxmlLoader.getController();
@@ -71,7 +71,7 @@ public class ShcParameterController {
      * Get whether a window is closed or not
      */
     public void getWindowStatus() {
-        if(windowChoiceBox.getValue() != null) {
+        if (windowChoiceBox.getValue() != null) {
             boolean isOpen = House.getWindowStatusOpen(windowChoiceBox.getValue());
             windowSetButton.setText(isOpen ? "Close" : "Open");
         }
@@ -81,7 +81,7 @@ public class ShcParameterController {
      * Open/Close a door
      */
     public void toggleDoorAction() {
-        if(doorsChoiceBox.getValue() != null) {
+        if (doorsChoiceBox.getValue() != null) {
             House.toggleDoor(doorsChoiceBox.getValue());
             windowAndDoorChoiceBoxInit();
             DashboardController dashboardController = App.fxmlLoader.getController();
@@ -93,18 +93,11 @@ public class ShcParameterController {
      * Get whether a door is closed or not
      */
     public void getDoorStatus() {
-        if(doorsChoiceBox.getValue() != null) {
+        if (doorsChoiceBox.getValue() != null) {
             boolean isOpen = House.getDoorStatus(doorsChoiceBox.getValue());
             doorSetButton.setText(isOpen ? "Close" : "Open");
         }
     }
-
-    public void openPermissionPage() throws IOException
-    {
-        Parent part = FXMLLoader.load(getClass().getResource("permissionCheck.fxml"));
-        Stage stage = new Stage();
-        Scene scene = new Scene(part);
-        stage.setScene(scene);
-        stage.show();
-    }
 }
+
+
